refactor: Separated tests into multiple files
