+++ packages/type-css-modules/tests/utils/css/get-class-names/edge-case-file-is-empty.test.js
@@ -1,0 +1,21 @@
+import { assert, loadFixture, test } from '@codemod-utils/tests';
+
+import { getClassNames } from '../../../../src/utils/css.js';
+import {
+  codemodOptions,
+  options,
+} from '../../../helpers/shared-test-setups/ember-app-flat.js';
+
+test('utils | css | get-class-names > edge case (file is empty)', function () {
+  const cssFile = '';
+
+  const inputProject = {
+    app: {
+      'styles.css': cssFile,
+    },
+  };
+
+  loadFixture(inputProject, codemodOptions);
+
+  assert.deepStrictEqual(getClassNames('app/styles.css', options), []);
+});
