--- packages/type-css-modules/tests/utils/css/get-css-declaration-file-paths.test.js
@@ -1,10 +1,0 @@
-import { assert, test } from '@codemod-utils/tests';
-
-import { getCssDeclarationFilePaths } from '../../../src/utils/css.js';
-import { options } from '../../helpers/shared-test-setups/ember-app-flat.js';
-
-test('utils | css | get-css-declaration-file-paths', function () {
-  assert.deepStrictEqual(getCssDeclarationFilePaths(options), [
-    'app/**/*.css.d.ts',
-  ]);
-});
