+++ packages/type-css-modules/tests/utils/css/get-declaration-file/edge-case-classNames-is-empty.test.js
@@ -1,0 +1,17 @@
+import { assert, test } from '@codemod-utils/tests';
+
+import { getDeclarationFile } from '../../../../src/utils/css.js';
+
+test('utils | css | get-declaration-file > edge case (classNames is empty)', function () {
+  const classNames = [];
+
+  assert.deepStrictEqual(
+    getDeclarationFile(classNames),
+    [
+      'declare const styles: Record<string, never>;',
+      '',
+      'export default styles;',
+      '',
+    ].join('\n'),
+  );
+});
